Introduce new website with css and html while using flexbox to arrange everything
This is the first section of my website. I focused on lining things up while adding some random filler conent for now while my primary focus with this website is the layout and styling. This section includes a header and title area with an embedded video on the page as well.