Introduce fourth section as well as the footer
This commit finishes the fourth and final secation of this webpage. I added a box that contains a call to action and sign up button. I rounded the edges a bit on the box and made everthing line up nicely inside. Also added a footer and centered the text while adding a good amount of padding as well.